Move IODB to separate package
