+++ global_segment_map_node/CMakeLists.txt
@@ -42,8 +42,6 @@ endif()
 
 cs_add_library(${PROJECT_NAME}_library
   src/controller.cpp
-  src/iodb_controller.cpp
-  src/sliding_window_controller.cpp
   src/feature_ros_tools.cpp
 )
 target_link_libraries(${PROJECT_NAME}_library ${catkin_LIBRARIES} ${approxmvbb_catkin_LIBRARIES})
@@ -52,16 +50,6 @@ cs_add_executable(${PROJECT_NAME}
   src/node.cpp
 )
 target_link_libraries(${PROJECT_NAME} ${PROJECT_NAME}_library)
-
-cs_add_executable(${PROJECT_NAME}_iodb
-  src/iodb_node.cpp
-)
-target_link_libraries(${PROJECT_NAME}_iodb ${PROJECT_NAME}_library)
-
-cs_add_executable(${PROJECT_NAME}_util
-  src/pose_to_tf_node.cpp
-)
-target_link_libraries(${PROJECT_NAME}_util ${PROJECT_NAME}_library)
 
 
 cs_install()
