+++ global_segment_map_node/src/feature_ros_tools.cpp
@@ -1,4 +1,4 @@
-#include "voxblox_gsm/feature_ros_tools.h"
+#include "global_segment_map_node/feature_ros_tools.h"
 
 namespace voxblox {
 namespace voxblox_gsm {
